extract LocaleStore, back to using map/object as valueTextMap
